Fix installments calculation by wrong discount application
Discounted for payment using the boleto was identified exchanging the boleto
for credit card, continuing to display the amount discounted in the options of
installments.

Checks the payment rule to display the amount discounted in the options of
installments.

Removes duplicated codes
Changes code of  place to organize the codes
Creates tests to cover the changes