tests: refactor configure e2e tests
Since this e2e suite is responsible to guarantee that module's
configurations can be saved and indicate errorns on it's xml
definitions this refactor intent to improve its execution time.

This suite is very slow to execute on CI (travis), so this refactor
merge all step definitions into a single scenario. It will avoid admin
authentication repeatedly turning this suite more faster to execute.