tests: create order (programmatically) by credit card
Create a new order by credit card using checkout transparent. To do
that, it defines a new environment variable `PAGARME_DEVELOPMENT=enabled` wich will be used to generate a valid `PagarMe\Sdk\Card` on `PagarMe_CreditCard_Model_Creditcard::generateCard` during the e2e tests.

The method `generateCard` is responsible to return an `PagarMe\Sdk\Card`
and, if the new environment variable is setted to `enabled` it means
that we're running magento in development environment, so, it'll return
a Card for tests.