checkout: made module compatible with firecheckout 4.3
To create credit card transactions, the module should be capable of
generate a [cardhash](https://docs.pagar.me/docs/overview-transacao#section-4-o-que-%C3%A9-um-card_hash).
This generation is executed before the user clicks on Place order button. This event is attached to the button via its parent,
that should be visible on the page, so. This use a generic parent id on
the page `#checkout-review-submit` that is present on firecheckout page.
And to don't break users who uses the magento's default checkout, the
event is attached to `#opc-review` too, that's the element visible on
the page.

Also, to make this work on firecheckout it was necessary to add our
javascript files to firecheckout page via xml configuration defining
when and which block those files should be inserted.