module setup: fix database migration
The platform uses install and upgrade scripts to setup the module and
database tables if the module needs to persist data on database.

This fixes the module install/upgrade scripts to avoid the table
recreation during the module upgrade. To do that, change the install
script (v3.5.5) to a upgrade script that will be executed based on the
major version `v3` that will be responsible to add a new column to
perists reference key value