creditcard: handle refused transactions during its creation
The module wasn't handling the transaction status if it's `refused`.
That's was causing a problem that the order was created as `processing`
and then through postback request its status was changed to `canceled`.

This PR fix the problem mentioned above setting the order as `pending`
if its related transaction is `refused` and then, via an observer,
create a comment history on admin panel describing the reason.