tests: implement workaround to set street lines config to 4
I don't like this implementation but, since that's the only way I found
to finish this task I'll use it. It sets the street lines config to 4
lines to allow admin/customers to buy on platform using Pagar.me's
module. This workaround sets the value programmatically but, to this
change be saved properly a press button on the admin page was needed.