order invoice: update totals when change products quantity
This commit recalculates the invoice's interest rate when the product
quantity is changed.

With this, if the seller choose to capture a order partially, the
interest rate will be calculated based on the products that going to be
invoiced.

about raw sql: our table pagarme_transaction has two primary keys, that
was blocking Zend query interface.