checkout: fix generate card hash flow
This PR changes the flow to generate a card hash.

Actually, it generates the card_hash when the customer clicks in place
order button.

With these modifications, the card_hash will be generated when the
customer input the card pieces of information, with de following rules:

1. all fields must have length >= 3
2. The pagarme credit card payment method must be selected
3. all fields must be validated with pagarme-js `cardValidations`

This way, our checkout doesn't depend on external elements (like the place order button)
to generate the card_hash and will be compatible with most external
modules.