order totals: avoid call on null
Reported on #431, an exception is throwed on order details page and
order creation process using another payment methods modules. It is
caused becaus a function call on null is performed outter of pagar.me's
module scope.

It checks if theres a referenced order before check if the order was
created using any of pagar.me's payment methods